requirements.txt
beaupy==3.5.4
python-dotenv==1.0.0
openai